+++ crates/core/config/src/workspace/hasher.rs
@@ -11,10 +11,21 @@ pub enum HasherOptimization {
     Performance,
 }
 
+#[derive(Clone, Debug, Default, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
+#[serde(rename_all = "kebab-case")]
+pub enum HasherWalkStrategy {
+    Glob,
+    #[default]
+    Vcs,
+}
+
 #[derive(Clone, Debug, Default, Deserialize, Eq, JsonSchema, PartialEq, Serialize, Validate)]
 #[schemars(default)]
 #[serde(default, rename_all = "camelCase")]
 pub struct HasherConfig {
     #[serde(skip_serializing_if = "is_default")]
     pub optimization: HasherOptimization,
+
+    #[serde(skip_serializing_if = "is_default")]
+    pub walk_strategy: HasherWalkStrategy,
 }