+++ packages/types/src/workspace-config.ts
@@ -4,6 +4,7 @@ export interface GeneratorConfig {
 
 export interface HasherConfig {
 	optimization: 'accuracy' | 'performance';
+	walkStrategy: 'glob' | 'vcs';
 }
 
 export interface NotifierConfig {